requirements.txt
# Dependências principais
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6

moviepy==1.0.3
numpy==1.26.0
imageio-ffmpeg==0.6.0
ffmpeg-python==0.2.0
opencv-python-headless==4.8.1.78
opencv-contrib-python-headless==4.8.1.78
